segmentation/requirements.txt
accelerate==1.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
jaxtyping==0.3.4
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psutil==7.2.0
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rad-dino==0.1.1
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tifffile==2025.5.10
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typer==0.21.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wadler_lindig==0.1.7
wandb>=0.16.0
timm